// src/types/generated/PromptComponent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PromptComponentType } from "./PromptComponentType";

export interface PromptComponent { id: number | null, versionId: number, componentType: PromptComponentType, name: string, content: string, variables: string | null, language: string, sortOrder: number, }